Refactor Query management
